+++ test/specs/unit/querySpec.js
@@ -510,6 +510,24 @@ describe("The Query Class", function() {
             expect(query.trigger).not.toHaveBeenCalled();
         });
 
+        it("Should not call _runConversation if the model is Conversation and pagedToEnd", function() {
+            spyOn(query, "_runConversation");
+            spyOn(query, "_runMessage");
+            spyOn(query, "_runAnnouncement");
+            spyOn(query, "_runIdentity");
+
+            spyOn(client, "_checkAndPurgeCache");
+            spyOn(query, "trigger");
+            query.data = [conversation];
+            query.pagedToEnd = true;
+
+            // Run
+            query._run();
+
+            // Posttest
+            expect(query._runConversation).not.toHaveBeenCalled();
+        });
+
         it("Should call _runMessage if the model is Message", function() {
             query.model = "Message";
             query.predicate = 'conversation.id = "fred"';
@@ -657,7 +675,7 @@ describe("The Query Class", function() {
             expect(query._processRunResults).toHaveBeenCalledWith(jasmine.objectContaining({
                 success: true,
                 data: [{id: "a"}, {id: "b"}]
-            }), "conversations?sort_by=created_at&page_size=36");
+            }), "conversations?sort_by=created_at&page_size=36", 36);
         });
     });
 
@@ -772,7 +790,7 @@ describe("The Query Class", function() {
             expect(query._processRunResults).toHaveBeenCalledWith(jasmine.objectContaining({
                 success: true,
                 data: [{id: "a"}, {id: "b"}]
-            }), "announcements?page_size=47");
+            }), "announcements?page_size=47", 47);
         });
     });
 
@@ -901,7 +919,7 @@ describe("The Query Class", function() {
             expect(query._processRunResults).toHaveBeenCalledWith(jasmine.objectContaining({
                 success: true,
                 data: [{id: "a"}, {id: "b"}]
-            }), "conversations/" + conversation.id.replace(/^layer\:\/\/\/conversations\//, "") + "/messages?page_size=47");
+            }), "conversations/" + conversation.id.replace(/^layer\:\/\/\/conversations\//, "") + "/messages?page_size=47", 47);
         });
 
         it("Should add lastMessage to the results", function() {
@@ -987,7 +1005,7 @@ describe("The Query Class", function() {
             expect(query._processRunResults).toHaveBeenCalledWith(jasmine.objectContaining({
                 success: true,
                 data: [{id: "a"}, {id: "b"}]
-            }), "identities?page_size=47");
+            }), "identities?page_size=47", 47);
         });
     });
 
@@ -1017,7 +1035,7 @@ describe("The Query Class", function() {
                 xhr: {
                     getResponseHeader: function() {return 6;},
                 }
-            }, requestUrl);
+            }, requestUrl, 10);
             expect(query.isFiring).toBe(false);
         });
 
@@ -1028,7 +1046,7 @@ describe("The Query Class", function() {
                 xhr: {
                     getResponseHeader: function() {return 6;},
                 }
-            }, requestUrl);
+            }, requestUrl, 10);
             expect(query.isFiring).toBe(false);
         });
 
@@ -1038,9 +1056,12 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {
-                    getResponseHeader: function() {return 6;},
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
                 }
-            }, requestUrl);
+            }, requestUrl, 10);
             expect(query._appendResults).toHaveBeenCalledWith({
                 success: true,
                 data: [{id: "a"}],
@@ -1048,7 +1069,7 @@ describe("The Query Class", function() {
             }, false);
         });
 
-        it("Should not call _run if reached the end of the server's results", function() {
+        it("Should not schedule _run if results are up to date", function() {
             spyOn(query, "_run");
             spyOn(query, "_appendResults");
             query.paginationWindow = 100;
@@ -1057,11 +1078,128 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {
-                    getResponseHeader: function() {return 6;},
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
                 }
-            }, requestUrl);
+            }, requestUrl, 10);
+            jasmine.clock().tick(10000);
             expect(query._run).not.toHaveBeenCalled();
             query.data = [];
+        });
+
+        it("Should schedule _run if results are not up to date", function() {
+            spyOn(query, "_run");
+            spyOn(query, "_appendResults");
+            query.paginationWindow = 100;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [{id: "a"}],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'true';
+                    }
+                }
+            }, requestUrl, 10);
+            jasmine.clock().tick(10000);
+            expect(query._run).toHaveBeenCalled();
+            query.data = [];
+        });
+
+        it("Should trigger server-syncing-state true if scheduling _run for the first time", function() {
+            spyOn(query, "_run");
+            spyOn(query, "trigger");
+            query.paginationWindow = 100;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 0;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'true';
+                     }
+                }
+            }, requestUrl, 10);
+            expect(query.trigger).toHaveBeenCalledWith('server-syncing-state', { syncing: true });
+        });
+
+        it("Should NOT trigger server-syncing-state true if scheduling _run for subsequent calls", function() {
+            spyOn(query, "_run");
+            spyOn(query, "trigger");
+            query._isServerSyncing = true;
+            query.paginationWindow = 100;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 0;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'true';
+                     }
+                }
+            }, requestUrl, 10);
+            expect(query.trigger).not.toHaveBeenCalledWith('server-syncing-state', jasmine.any(Object));
+        });
+
+        it("Should trigger server-syncing-state false if sufficient data", function() {
+            spyOn(query, "_run");
+            spyOn(query, "trigger");
+            query._isServerSyncing = true;
+            query.paginationWindow = 4;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 0;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'true';
+                     }
+                }
+            }, requestUrl, 10);
+            expect(query.trigger).toHaveBeenCalledWith('server-syncing-state', { syncing: false });
+        });
+
+        it("Should trigger server-syncing-state false if server is done syncing", function() {
+            spyOn(query, "_run");
+            spyOn(query, "trigger");
+            query._isServerSyncing = true;
+            query.paginationWindow = 100;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 0;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
+            }, requestUrl, 10);
+            expect(query.trigger).toHaveBeenCalledWith('server-syncing-state', { syncing: false });
+        });
+
+        it("Should not trigger server-syncing-state false if server is not syncing", function() {
+            spyOn(query, "_run");
+            spyOn(query, "trigger");
+            query.paginationWindow = 100;
+            query.data = [message, message, message, message, message, message, message];
+            query._processRunResults({
+                success: true,
+                data: [],
+                xhr: {
+                     getResponseHeader: function(name) {
+                         if (name == 'Layer-Count') return 0;
+                         if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                     }
+                }
+            }, requestUrl, 10);
+            expect(query.trigger).not.toHaveBeenCalledWith('server-syncing-state', { syncing: false });
         });
 
         it("Should not call _appendResults if request is not the most recent request", function() {
@@ -1071,7 +1209,7 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {}
-            }, 'joe');
+            }, 'joe', 10);
             expect(query._appendResults).not.toHaveBeenCalled();
         });
 
@@ -1082,7 +1220,7 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {}
-            }, 'joe');
+            }, 'joe', 10);
             expect(query._appendResults).not.toHaveBeenCalled();
 
             // Cleanup
@@ -1097,7 +1235,7 @@ describe("The Query Class", function() {
                 success: true,
                 data: [{id: "a"}],
                 xhr: {}
-            }, 'joe');
+            }, 'joe', 10);
             expect(query.isFiring).toBe(true);
             expect(query._firingRequest).toEqual('fred');
         });
@@ -1109,8 +1247,30 @@ describe("The Query Class", function() {
                 xhr: {
                     getResponseHeader: function() {return 6;},
                 }
-            }, requestUrl);
+            }, requestUrl, 10);
             expect(query.totalSize).toEqual(6);
+        });
+
+        it("Should not set pagedToEnd if the number of requested results is returned", function() {
+             query._processRunResults({
+                success: true,
+                data: [responses.message1, responses.message1, responses.message1, responses.message1],
+                xhr: {
+                    getResponseHeader: function() {return 6;},
+                }
+            }, requestUrl, 4);
+            expect(query.pagedToEnd).toBe(false);
+        });
+
+        it("Should set pagedToEnd if less than the number of requested results is returned", function() {
+             query._processRunResults({
+                success: true,
+                data: [responses.message1, responses.message1, responses.message1],
+                xhr: {
+                    getResponseHeader: function() {return 6;},
+                }
+            }, requestUrl, 4);
+            expect(query.pagedToEnd).toBe(true);
         });
     });
 
@@ -1131,18 +1291,42 @@ describe("The Query Class", function() {
         it("Should register new results", function() {
             spyOn(client, "_createObject");
             spyOn(client, "_getObject").and.returnValue(conversation);
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+            });
             expect(client._createObject).toHaveBeenCalledWith(responses.conversation2);
         });
 
         it("Should update _nextDBFromId if there is data from DB", function() {
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]}, true);
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+            }, true);
             expect(query._nextDBFromId).toEqual(responses.conversation2.id);
             expect(query._nextServerFromId).toEqual('');
         });
 
         it("Should update _nextServerFromId if there is data from Server", function() {
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]}, false);
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+            }, false);
             expect(query._nextServerFromId).toEqual(responses.conversation2.id);
             expect(query._nextDBFromId).toEqual('');
         });
@@ -1160,7 +1344,15 @@ describe("The Query Class", function() {
             conversation.createdAt.setHours(conversation.createdAt.getHours() + 1);
 
             // Run
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+            });
 
             // Posttest
             expect(query.data).not.toBe(oldData);
@@ -1179,7 +1371,15 @@ describe("The Query Class", function() {
             var oldData = query.data = [conversation];
 
             // Run
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+            });
 
             // Posttest
             expect(query.data).toBe(oldData);
@@ -1188,7 +1388,15 @@ describe("The Query Class", function() {
 
         it("Should put objects rather than instances if dataType is object", function() {
             query.dataType = "object";
-            query._appendResults({data: [JSON.parse(JSON.stringify(responses.conversation2))]});
+            query._appendResults({
+                data: [JSON.parse(JSON.stringify(responses.conversation2))],
+                xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+            });
             expect(query.data[0] instanceof layer.Conversation).toBe(false);
         });
 
@@ -1208,7 +1416,15 @@ describe("The Query Class", function() {
           }).and.returnValue(1);
 
           // Run
-          query._appendResults({data: [c3]});
+          query._appendResults({
+              data: [c3],
+              xhr: {
+                getResponseHeader: function(name) {
+                    if (name == 'Layout-Count') return 6;
+                    if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                }
+            }
+          });
 
           // Posttest
           expect(query._getInsertConversationIndex).toHaveBeenCalled();
@@ -1232,7 +1448,15 @@ describe("The Query Class", function() {
           }).and.returnValue(1);
 
           // Run
-          query._appendResults({data: [m3]});
+          query._appendResults({
+              data: [m3],
+              xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+          });
 
           // Posttest
           expect(query._getInsertMessageIndex).toHaveBeenCalled();
@@ -1261,7 +1485,15 @@ describe("The Query Class", function() {
           query.model = layer.Query.Identity;
 
           // Run
-          query._appendResults({data: [i3]});
+          query._appendResults({
+              data: [i3],
+              xhr: {
+                    getResponseHeader: function(name) {
+                        if (name == 'Layout-Count') return 6;
+                        if (name == 'Layer-Conversation-Is-Syncing') return 'false';
+                    }
+                }
+          });
 
           // Posttest
           expect(query.data).toEqual([i1.toObject(), i2.toObject(), i3.toObject()]);